requirements.txt
djangorestframework==3.14.0 
Django==4.1.2
Pillow==9.3
sqlparse==0.4.3
asgiref==3.5.2
psycopg2==2.8.6